requirements.txt
numpy~=2.0.1
h5py~=3.12.1
pillow~=11.0.0
matplotlib~=3.9.2
scipy~=1.14.1
tensorly~=0.8.0